Remove async support
Remove async support for now as the current code is not running asynchronously anyway. Sindri will most likely run on a single core machine so the only point would of async is to either
a) interrupt long-running software computations or
b) hand over computations to dedicated hardware engines.

The first case requires the crypto crates that we use to support async as well. As of now, none of the crates we use does this. Whether the second case will be implemented using async is not clear yet. Until then, we continue non-async.