Remove memory allocation and replace with slices
This change completely removes the memory management from Heimlig,
instead relying on slices provided in the requests for both input and
output parameters.

As part of this change, the `Channel` concept was replaced with a
`Source` + `Sink` abstraction which lends itself better to a future use
of async `Stream` and `Sink` from futures.