Provide blanket implementation of "secure" methods for KeyStore
Splits the `KeyStore` trait into a base "insecure" version, which
is intended to provide the raw access to the keystore without checks for
permissions, and provides a blanket implementation of `KeyStore` for any
type implementing `InsecureKeyStore`. This way, users can implement
their own `KeyStore` if they want to fully control the checks, or they
can provide an `InsecureKeyStore` and get for free the basic permission
and key type checks. The `KeyStore` trait re-exposes the insecure
methods so that workers can make internal use of them.