Refactor unsafe code in raw_jobs.rs for better memory safety
A potential issue was present where the check to the pointed to memory
was done independently to the cast to the Rust type. The check was meant
to protect from casting memory in an invalid state, but did not protect
from the memory being modified in between the check and the cast.

This new version creates a copy on the stack _first_. This way, we can
safely validate it and trust our own memory to still be valid when we
call `assume_init()`.